rename configs_class to config_class
