+++ datashuttle/configs/canonical_configs.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, List
 
 if TYPE_CHECKING:
-    from datashuttle.configs.configs_class import Configs
+    from datashuttle.configs.config_class import Configs
 
 from pathlib import Path
 from typing import Literal, Union, get_args, get_origin
