+++ datashuttle/utils/data_transfer.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import List, Optional, Union
 
-from datashuttle.configs.configs_class import Configs
+from datashuttle.configs.config_class import Configs
 
 from . import directories, formatting, rclone, utils
