+++ datashuttle/utils/ssh.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from datashuttle.configs.configs_class import Configs
+    from datashuttle.configs.config_class import Configs
 
 import fnmatch
 import getpass
